🔨 Refactor helpers to move last_date_accessed and last_date_published functions to catalog_helpers
