+++ etl/helpers.py
@@ -2,7 +2,6 @@
 #  helpers.py
 #  etl
 #
-import datetime as dt
 import re
 import time
 from functools import cache
@@ -809,23 +808,3 @@ def create_explorer(
         explorer.df_columns = df_columns
 
     return explorer
-
-
-def last_date_accessed(tb: Table) -> str:
-    """Get maximum date_accessed from all origins in the table and display it in a specific format.
-
-    Usage:
-        create_dataset(..., yaml_params={"date_accessed": last_date_accessed(tb)})
-    """
-    date_accessed = max([origin.date_accessed for col in tb.columns for origin in tb[col].m.origins])
-    return dt.datetime.strptime(date_accessed, "%Y-%m-%d").strftime("%d %B %Y")
-
-
-def last_date_published(tb: Table) -> str:
-    """Get maximum date_published from all origins in the table and display it in a specific format.
-
-    Usage:
-        create_dataset(..., yaml_params={"date_published": last_date_published(tb)})
-    """
-    date_published = max([origin.date_published for col in tb.columns for origin in tb[col].m.origins])
-    return dt.datetime.strptime(date_published, "%Y-%m-%d").strftime("%d %B %Y")
