+++ etl/steps/data/garden/artificial_intelligence/2025-01-08/epoch_compute_intensive_countries.py
@@ -2,8 +2,9 @@
 
 import shared as sh
 
+from etl.catalog_helpers import last_date_accessed
 from etl.data_helpers import geo
-from etl.helpers import PathFinder, create_dataset, last_date_accessed
+from etl.helpers import PathFinder, create_dataset
 
 # Get paths and naming conventions for current step.
 paths = PathFinder(__file__)
