+++ etl/steps/data/garden/artificial_intelligence/2025-02-04/epoch_compute_intensive_domain.py
@@ -2,7 +2,8 @@
 
 import shared as sh
 
-from etl.helpers import PathFinder, create_dataset, last_date_accessed
+from etl.catalog_helpers import last_date_accessed
+from etl.helpers import PathFinder, create_dataset
 
 # Get paths and naming conventions for current step.
 paths = PathFinder(__file__)
