+++ etl/steps/data/garden/climate/latest/weekly_wildfires.py
@@ -3,8 +3,9 @@
 import owid.catalog.processing as pr
 import pandas as pd
 
+from etl.catalog_helpers import last_date_accessed
 from etl.data_helpers import geo
-from etl.helpers import PathFinder, create_dataset, last_date_accessed
+from etl.helpers import PathFinder, create_dataset
 
 # Get paths and naming conventions for current step.
 paths = PathFinder(__file__)
