+++ etl/steps/data/grapher/climate/2023-12-20/surface_temperature_annual_average.py
@@ -3,7 +3,8 @@
 import numpy as np
 import owid.catalog.processing as pr
 
-from etl.helpers import PathFinder, create_dataset, last_date_accessed
+from etl.catalog_helpers import last_date_accessed
+from etl.helpers import PathFinder, create_dataset
 
 # Get paths and naming conventions for current step.
 paths = PathFinder(__file__)
